techpack: audio: q6voice: Force AVCS CVP version 1 on legacy SoC
Force the AVCS version to CVP_VERSION_1 for all of the AVCS
services on sdm636: It cannot execute the Q6 Service Version
command on it's Q6 DSP because of older firmware not supporting it.
We know that for this SoC the version is always 1, hence
forcing it is totally safe.

Change-Id: I81e33a1456866e11af33320c74995a4e6f4d5a12
Signed-off-by: Kunmun <kunmun.devroms@gmail.com>