tracing: move function tracer options to Kconfig
Move function tracer options to Kconfig to make it easier to add
new methods for generating __mcount_loc, and to make the options
available also when building kernel modules.

Note that FTRACE_MCOUNT_USE_* options are updated on rebuild and
therefore, work even if the .config was generated in a different
environment.

Acked-by: Steven Rostedt (VMware) <rostedt@goodmis.org>
Link: https://lore.kernel.org/r/20201211184633.3213045-2-samitolvanen@google.com
Change-Id: I3a7769d176500bbb687b034a6cc0996daa3cb856
Signed-off-by: Kunmun <kunmun.devroms@gmail.com>