kbuild: remove top-level built-in.a
The symbol table in the final archive is unneeded; the linker does not
require the symbol table after the --whole-archive option. Every object
file in the archive is included in the link anyway.

Pass thin archives from subdirectories directly to the linker, and
remove the final archiving step.

Fix up the document and comments as well.

Acked-by: Nicholas Piggin <npiggin@gmail.com>
Change-Id: I0984ff27d776ede5f7c8debf8b58105496117bc7
Signed-off-by: Kunmun <kunmun.devroms@gmail.com>