msm/sde/rotator: Remove unneeded PM QoS requests
When the rotator is actually used (still an unsolved question in
computer science), these PM QoS requests block some CPUs in the LITTLE
cluster from entering deep idle because the driver assumes that display
rotating work occurs on a hardcoded set of CPUs, which is false. We
already have the IRQ PM QoS machinery for display rendering operations
that actually matter, so this cruft is unneeded.

Change-Id: I7ee9e8aa4c641d5db5a6a42cb7c478e7f53f2e48
Signed-off-by: Kunmun <kunmun.devroms@gmail.com>