arm64: vdso: remove -nostdlib compiler flag
The -nostdlib option requests the compiler to not use the standard
system startup files or libraries when linking. It is effective only
when $(CC) is used as a linker driver.

Since commit 691efbe ("arm64: vdso: use $(LD) instead of $(CC)
to link VDSO"), $(LD) is directly used, hence -nostdlib is unneeded.

Link: https://lore.kernel.org/r/20211107161802.323125-1-masahiroy@kernel.org
Change-Id: Iacfbc19e2e33802b8a01352866a6256e7ff51be4
Signed-off-by: Kunmun <kunmun.devroms@gmail.com>